Adding hal.device.query/iree_hal_device_query_*. (#5421)
This enables simple runtime configuration queries from HAL devices.
The exact set of things that can be queried will vary and this is just
the plumbing. As each key supported puts pressure on version compatibility
and testing we'll want to add them with care (more notes in the .td).

I32 only for right now; we can add additional types if needed though they
may require additional VM support and we want to avoid conditional bitwidth
usage (like I64 which may not be available everywhere).