Add attribute to capture lowering configuration. (#5842)
Current approach to capturing lowering configuration is to use a C++
object. This was added temporarily, but has become load-bearing. A
(slightly) better approach is to just compute the configuration and
add it as an attribute on the operation. All the lowering through
tiling and vectorization then uses this attribute.
The use of the attribute is plumbed through on the CPU side at this
time.

This is the first step towards using dynamic pass pipeline usage.