Moving HAL matchers to use hal.device.query. This gives us a single runtime call to perform both parametric value queries as well as pattern-matching queries. There's a great deal of work still to do here but this unblocks iteration on the layering.
As part of this hal.device.query gained a namespace that can be used to
disambiguate queries that may route to built-in hal features (like
hal.device.id) or backend-specific ones (vulkan.whatever). The interface
is intentionally vague about how these are treated so that we can adapt
to all types without plumbing through additional query-like methods from
compiler to runtime.