Make Dispatch region formation handle `LinalgExtOp`s that implement `TiledOpInterface`. (#6540)
Using the pattern to tile + distribute operations that implement the
TiledOpInterface, this change extends dispatch region creation logic
to handle LinalgExpOp that implement the tiled op interface.

Also a few clean ups

Add utility functions to interpret the root and fusion attributes to
make the intent clearer.
Modify the logic that decides the loops to be partitioned to be more
configurable
Make the analysis to find tied operands of the dispatch region run
after destructive updates are resolved to use
flow.dispatch.tensor.load/store. This seems to be a bug, that
would disallow having any LinalgOp not have tied operands either.
Remove the use of AffineMinSCFCanonicalizationPattern. It seems to
hit an assertion (see Issue Segfault during application of AffineMinSCFCanonicalizationPattern #6520), also since all loop bounds are
dynamic, there is no reason to apply this canonicalization.