Some cleanup to docker images (#7406)
At this point all of our builds use Python. With
https://github.com/google/iree/pull/7347 the last holdouts are growing a
dep on yaml to be able to run traces. So I squashed `cmake-python` into
`cmake` and shortened all the dependents by dropping `python` from their
name.

Some other cleanups, some of which were unfortunately necessary to be able to rebuild our images. Hooray unreproducible builds!
1. The rbe-toolchain image was never actually updated to bump the swiftshader commit in https://github.com/google/iree/pull/7117 and
https://github.com/google/iree/pull/6859, and the newer commits don't
actually build there so I had to restore the commit in the Dockerfile to
the one that has actually always been the one the image is using. See
https://github.com/google/iree/issues/7419.
2. The place we were fetching the Vulkan SDK from is no longer
available, so I had to rebundle it and stash it in our GCS. Upgrading
Vulkan is a huge undertaking that I'm not going to undertake right now.
3. Deleted the `cmake-bazel-frontends-android` image, which is
unused.