Move `LoweringConfigAttr` and `TranslationInfoAttr` to `AttrDef`. (#7385)
Previously these were defined using StructAttr which is limited in
its capabilities. Moving them to AttrDef allows this to be more
robust and verifiable. Also allows specification of custom builder
methods that remove a lot of the bioler plate.

In addition to the above, this change also adds a new attribute,
CompilationInfoAttr. This is the attribute that when added to
operations like linalg.matmul/linalg.*conv* (or their mhlo/tosa
couterparts) use the information from the specified attribute instead
of the default heuristics. This allows an external search to search
for better values than the defaults used. While there are some
preliminary verifies in place, additional verifiers are needed to
ensure that the user-provided attribute has all the information needed
(and in a consistent manner) to achieve correct compilation.

In addition, all these attributes are moved from the HAL dialect into
a new IREECodegenDialect added to compiler/Codegen/Dialect.

This change also brings in some code from D111594 to allow for this to
land without having to wait for that patch to land.