Refactor CMake rules for connecting to TF binaries (#7479)
This makes the imported TF binaries more uniform with our other binaries
in terms of their naming and the convenience aliases. It also factors
some common logic into a function.

This is only one option for changing the way the binaries are
referenced. We could also use unqualified name (since we set it up so
our binaries are globally unique and available by their basename). I
don't think we ever came up with a consistent rule for that, other than
that we shouldn't use the underscore-mangled name that only exists 
because CMake has weird restrictions on some target names (but not
aliases).