enable asserts in a majority of CMake CI builds (#7482)
Bazel builds already had asserts enabled, so this is only about CMake builds.

It's generally useful to build CI with asserts, but we wouldn't want to build only asserts-builds at all: after all, the notion that asserts are only actually assertions and don't play a part in the program's actual function, is left entirely up to the programmer, and we want to catch mistakes at that level e.g. instances of assert(is_status_ok(important_file_io()) expanding to nothing in non-asserts builds.

So this PR adds IREE_ENABLE_ASSERTIONS=ON to some common build.sh scripts used for CI, but intentionally not all of them. For example, the -asan CI uses its own build.sh file and we are intentionally not adding asserts there just so we keep some non-asserts coverage (this isn't embodying an opinion about whether to conflate asserts and asan in the same builds or to keep them separate! it's just what was quick to do here).

Fixes Issue #7320 .