Implement benchmarks from TFLite flatbuffers (#7493)
These will currently get skipped if the TFLite importer is not enabled,
which is the case in the current benchmark pipeline, so it should have
no affect on that. A few more things are needed before that can be
turned on. In that case, I'll actually replace the current benchmarks
instead of adding a new one, since these are actually coming from the
exact same source.

Tested: Locally built and ran android benchmarks.