Avoid using Bazel in the Buildkite Android pipeline (#7541)
Turns out the Buildkite Android build was using a script under the
Kokoro directory. https://github.com/google/iree/pull/7494 switched
that script to build the TFLite importer with Bazel to allow verifying
that we can build benchmark artifacts for Android (the least-bad option
available at the moment, see
https://github.com/google/iree/pull/7494#issuecomment-959710770). We
don't want to be doing that as part of the Buildkite pipeline though.
So we pull the Bazel stuff into the Kokoro script and make the
Buildkite pipeline not use it.