Enable tiling of linalg.pad_tensor operations. (#7591)
Forwarding the implementation of TiledOpInterface to methods in
linalg.pad_tensor that implement the TilingInterface in MLIR
allows IREE to pick up the tiling transformations that exist on this
operation. This PR just checks that the pad tensor operation is tiled
during dispatch region formation. It isnt connected end-to-end yet
till the effects downstream can be evaluated.