Changing tflite binding generation to use hal.tensor.cast. (#7661)
This allows for the buffers to be directly specified (just like we
specify buffer_views in the normal IREE bindings) and lets us drop all
of the shapex dialect ops from this layer.

Previously this was two passes but it relied on some shady assumptions
about when and how shapes were propagated. Now as a single pass all global
expanded shape dimensions, the ops tying them to values, and the logic
for query/update are explicitly specified and robust to further
transformation.