Only check submodule status once per git commit. (#7671)
Fixes https://github.com/google/iree/issues/6575

`git submodule status` is slow, so running it every cmake configure is annoying. This caches the latest git commit that the submodule init check ran successfully at. The check is also a no-op if git or python are not installed, or the source tree is not a git repository.