Splitting librt into two builtins and changing linking style. (#7721)
This is based on the work done in #7561 to better link the bitcode in the
compiler. Instead of generating object files to pass to the linker tool
this now links the bitcode modules directly together before running the
LLVM optimization passes. This gets us LTO-like behavior and makes the
bitcode equivalent to any other LLVM IR we could generate in the compiler.

As part of this the IREE-specific portion has been reworked to compile
both standalone and as part of a host binary. This lets us write tests
and benchmarks for the functions in there and debug them independently
of the compiler.

The new setup will also allow us to specialize for particular archs if we
really needed. For example, if we added MMA-like CPU primitives we could
build them out in here. Each of the Builtins/ files can make a decision
as to which bitcode file to link and what options to set based on the
target configuration being compiled (selecting SIMD variants, etc).

There's still some stuff to work out but this drops our embedded ELFs by
13KB (on x64) and will allow us to add more musl/runtime functions
without bloating binaries. I think getting further will require us to generate
a linker script.

This also prepares us for removing the linker requirement when generating
static libraries as we can include all of the code a user needs in the single
static object we emit.