Set CPUs to max frequency before running benchmarks (#7728)
The aim is that this will reduce variance in benchmarks. Notably, this
reduces our ability to capture important real-world scenarios, like
running across threads with different clock speeds. We are deliberately
making a tradeoff right now to get variance under control and we may
revisit this in the future. Pinning to max clock speed is safe and
consistent only because the benchmark phones are on cooling plates so
thermal throttling won't be a problem.

Only the Pixel phone is pinned because we don't have the Samsung phones
rooted (because Samsung makes this a PITA and doing so caused the phone
to randomly reboot all the time).