Allow to turn of error on uninitialized submodules (#7911)
Makes the error on uninitialized submodules optional. With this,
external projects that use IREE are no longer forced to initialize all
submodules. E.g. if a project only uses the runtime and does not reply
on the TensorFlow integration, these submodules don't have to be cloned
at all. This is especially convenient in CI pipelines.