Ensuring buffer deallocation always routes through allocators. (#8082)
If the HAL module was the last retainer of a buffer (such as those
whose last use was in a command buffer) it was bypassing the device
allocator and directly destroying the buffer. CUDA (and ROCM) require
the allocator to free their buffer memory and this resulted in them
leaking device memory even if it appeared that memory was freed from
our side.