Stripping assert ops from within flow.executables. (#8101)
Frontends can introduce these and they don't work with our async
execution model. We could support them in the future using the same
kind of mechanism as #8015: pass in a system buffer to dispatches that
could assert and atomic rmw a failure flag we can then lazily check on
the host to indicate device loss.