Use the real lit tool for lit testing (#8166)
For historical reasons, we've used a custom shell script to run our lit
tests. We've made it less bad over time, but it's time to just use the
real thing. This gets the real lit tool running under Bazel (and CMake,
but that's less hard).

I hope to upstream some of this Bazel stuff, but LLVM's lit setup is
way more complicated, so checking this in here for now.

Fixes https://github.com/google/iree/issues/818