Generate arm intrinsics for MMT4D (#7751)
This passes all the e2e matmul tests (I added one specifically for the
MMT4D shape, which I think was missing test coverage). The option to
use intrinsics vs asm is added as a global flag to be used for
*development* as we explore these two different paths. It defaults to
using asm, since that version is actually producing the thing we want
right now. Registration of Arm Neon dialects and passes is added
unconditionally in a few places. I can look into making that
conditional if that seems worthwhile.

The generated asm is also not great. The LLVM shuffle + vdotq_s32
intrinsic is not getting switched into the version that uses lanes
directly. I wonder if the issue has to do with no way to express poison
at this level.

Here's a gist showing a lowering down to asm of the lit test here:
https://gist.github.com/GMNGeoffrey/02509944091560adf8150ceb2445cb27

In contrast to inline asm:
https://gist.github.com/GMNGeoffrey/06c4bb92708f1d3d2bc173a4ceafe5dd

Regardless, this isn't in a production path, so I think we can optimize
later.