Add special case of mul(float(bool), float) (#8396)
Multiplication of cast from bool is a common method for conditional
subtraction. Unfortunately it is expect that mul(inf, 0.0) -> 0.0, which is
not conformant with the IEEE floating point standard.

Added a special case to search for this substitution and perform a select
instead. It won't catch all cases but it should help track them down.