Explicitly find linker in either environment or own build/install dir (#8670)
While working on the next commit in this series, on UnixLinkerTool which handles both Apple and general-Unix cases, I realized that we had a bit too much magic nestled in findToolInEnvironment: having it search both our own executable dir and the environment made for confusing call sites where it wasn't clear when we're interested in one or the other. This would be useful if we actually needed to search both, but in practice, it seems that each call site knows which or the two cases it falls in (own executable dir or environment) so this makes code more explicit in this respect and allows the following commit to be less error-prone.

I had to reinstate searching the environment in EmbeddedLinkerTool because two of the CI builds failed to find the linker without that. I guess it's good that we discovered that.

I dropped some explicit inline storage size hints in SmallVector's -- rewrote SmallVector<std::string, 4> to just SmallVector<std::string> -- because otherwise i had to propagate that (can't pass a Smallvector<T, N> to a function that expects a SmallVector<T, K>) and that felt like premature optimization.