UnixLinker: trim toolnames (Apple) and generate shared lib (non-Apple) (#8673)
On Apple, now that it's clear that we're only searching the environment for a system linker, it's also clear that we only need to search for ld.
Outside of Apple, similarly to the Android situation before PR #8665, we weren't so far generating a shared object. Doing so makes us explicitly require LLD so we trim the toolnames list accordingly.