Fix scalarize_dyn_dims composition. (#8725)
* Fix scalarize_dyn_dims composition.

Previously, tiling with the scalarize_dyn_dims would systematically crash because it
expected `0` loops to be produced (i.e. because the tile size must be empty).
In reality, this will always produce as many loops as there are dynamic dims in the op.

Change the behavior to only return the op and always return `0` loops.
In the future this should become its own transformation, separated from tiling.

* Introduce a ScalarizeOp and remove the option from TilingOp

This allows for better separation of concerns.