Swap Stage 5 and Stage 6 in Codegen Driver (#8722)
Shape cast operations are usually no-ops that canonicalize away at some
point in the pipeline. However, if they are not, the codegen driver lowers
them to an inefficient sequence of insert/extract ops. This lowering is
enabled at stage 5 *and later ones* by default. Unfortunately, the AVX2 lowering
for transpose operations (stage 6) generates shape cast operations that
could be canonicalized away but are immediately lowered to the insert/extract
sequence before canonicalization can happen.

This patch swaps stage 5 and 6 to address this problem.