Fixes configurations for matvec and dot.  (#8775)
In the past, we only considered contraction ops that have at least
three loops. The PR accounts matvec and dot cases.

For now, set the limit to having a single reduction loop and it's
expected to be the innermost loop. This condition can be relaxed if
there are needs.

Fixes https://github.com/google/iree/issues/8713